move standardisation methods to individual files
